common: Move CurrentPic to new module
Need to figure out how to replace the `GtkPixbuf` with something not
GUI-centric.
