+++ gui.c
@@ -11,6 +11,7 @@
 #include "listen.h"
 #include "modespec.h"
 #include "pcm.h"
+#include "pic.h"
 
 GuiObjs      gui;
