+++ pic.c
@@ -1,0 +1,3 @@
+#include "pic.h"
+
+PicMeta      CurrentPic;
