+++ video.c
@@ -12,6 +12,7 @@
 #include "gui.h"
 #include "modespec.h"
 #include "pcm.h"
+#include "pic.h"
 #include "video.h"
 
 /* Demodulate the video signal & store all kinds of stuff for later stages
